weaviate/requirements.txt
Flask==3.0.2
weaviate-client>=4.4.4
python-dotenv
langchain-openai>=0.0.8
langchain-weaviate>=0.0.3
langchain-core>=0.1.40